pyproject.toml
[project]
name = "github-grader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.4"
dependencies = [
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
